% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_indices.R
\name{append_with_cat}
\alias{append_with_cat}
\title{append_with_cat -}
\usage{
append_with_cat(output, numeric_inputs, string_data)
}
\arguments{
\item{output}{- a vector of Y (target) values}

\item{numeric_inputs}{- a matrix of X values - i.e., all numeric inputs}

\item{string_data}{- all non-numeric columns in the data set}
}
\value{
all_inputs      - a matrix containing all numeric and all transformed input variables

sorted_X        - the order of categories in increasing average Y for each non-numeric input variables
}
\description{
Appends numeric inputs with categorical variables which are transformed into numeric ones.
}
\examples{

data            <- data.frame (Y  = c(10,20,40,30,60,10),
                               X1 = c(1,2,2,2,1,1),
                               X2 = c(0.6,0.8,0.9,0.4,0.2,0.3)
                               X3 = c("Y", "N", "N", "N", "Y", "Y")
AWC             <- cat_transform(data[,1], data[, 2:3], data[,4])
all_inputs      <- AWSC$all_inputs
head(all_inputs)
}
