% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Significance.R
\name{bin_data_1D}
\alias{bin_data_1D}
\title{bin_data_1D -}
\usage{
bin_data_1D(X, Y, n_bins_default)
}
\arguments{
\item{X}{- a vector of X values}

\item{Y}{- a vector of Y values}

\item{n_bins_default}{- default number of bins}
}
\value{
bin_avg       - averages of Y values in bins of X

bin_count     - number of Y values in bins of X
}
\description{
bin_data_1D computes averages and counts of Y in bins of X. It uses the magic_binning function.
}
\examples{

X              <- c(10, 20, 30, 15, 8, 40, 60, 10, 17, 4)
Y              <- c(167, 82, 75, 134, 186, 51, 17, 167, 86, 198)
n_bins_default <- 3
BD1D <- bin_data_1D(X, Y, )
}
