% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Significance.R
\name{bin_data_2D}
\alias{bin_data_2D}
\title{bin_data_2D -}
\usage{
bin_data_2D(Xi, Xj, Y, n_bins_default)
}
\arguments{
\item{Xi}{- a vector of Xi values}

\item{Xj}{- a vector of Xj values}

\item{Y}{- a vector of Y values}

\item{n_bins_default}{- default number of bins}
}
\value{
bin_avg_ij_string    - an array of average Y values in 2D bins of Xi and Xj

bin_count_ij_string  - an array of corresponding count of Y values

bin_avg_i            - an array of average Y values in 1D bins of Xi

bin_count_i          - an array of corresponding count of Y values

bin_avg_j            - an array of average Y values in 1D bins of Xj

bin_count_j          - an array of corresponding count of Y values
}
\description{
bin_data_2D computes averages and counts of Y in bins of XiXj, Xi, and Xj. The function uses the magic_binning function.
}
\examples{
df <- data.frame(Y  <- c(167, 82, 75, 134, 186, 51, 17, 167, 86, 198),
                 X1 <- c(10, 20, 30, 15, 8, 40, 60, 10, 17, 4),
                 X2 <- c(4,9,8,7,6,17,6,7,8,3))
n_bins_default <- 4
BD2D           <- bin_data_2D(df[,2], df[,3], df[,1], n_bins_default)
}
