% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Significance.R
\name{cat_transform}
\alias{cat_transform}
\title{cat_transform -}
\usage{
cat_transform(Y, X)
}
\arguments{
\item{Y}{- a vector of Y (target) values}

\item{X}{- a vector of X values}
}
\value{
sortedX  - categories of X sorted by increasing average Y

indexMap - an array of the same size as X with assigned numeric
                   indices corresponding to original categorical values
}
\description{
Transforms categorical variable into numeric, by assigning indices in
 accordance with average Y values in each category.
}
\examples{
-

Y               <- c(10,20,40,30,60,10)
X               <- c(1,2,2,2,1,1)
CT              <- cat_transform(Y, X)
sortedX         <- CT$sortedX
indexMap        <- CT$indexMap
print(sortedX)
print(indexMap)
}
