% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Significance.R
\name{magic_binning}
\alias{magic_binning}
\title{magic_binning -}
\usage{
magic_binning(X, n_bins_default)
}
\arguments{
\item{X}{- a numeric vector of values}

\item{n_bins_default}{- default number of bins}
}
\value{
bin_idx         - an array with bin indices of the same length as X

n_bins_out      - resulting number of bins (might be less than default if many NaNs or same values)
}
\description{
a function that bins values following a set of rules. n_bins_default is obtained using the number_of_bins function.
}
\details{
The following binning rules apply:
1. The size of each bin should be no less than length(X)/n_bins.
2. The same values of X should be allocated to a single bin.
3. NA values should get the 0 bin index.
}
\examples{
X <- c(NA, NA, 6, 6, 6, 6, 6, 6, 100, 200, 2, 1)
n_bins_default = 6
MB <- magic_binning(X, n_bins_default)
}
