% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_indices.R
\name{number_of_bins}
\alias{number_of_bins}
\title{number_of_bins -}
\usage{
number_of_bins(N_runs, N_factors)
}
\arguments{
\item{N_runs}{- the total number of observations}

\item{N_factors}{- number of inputs variables}
}
\value{
N_bins_foe - number of bins for first-order effect calculation

N_bins_soe - number of bins for second-order effect calculation
}
\description{
number_of_bins defines the optimal number of bins for first-order and
second-order significance indices calculation.
}
\examples{

data            <- data.frame (Y  = c(10,20,40,30,60,10),
                               X1 = c(1,2,2,2,1,1),
                               X2 = c(0.6,0.8,0.9,0.4,0.2,0.3),
                               X3 = c("Y", "N", "N", "N", "Y", "Y"))
N_runs          <- nrow(data)
N_factors       <- ncol(data) - 1 # Assuming one column of the data is for the target variable, i.e., Y
NB              <- number_of_bins(N_runs, N_factors)
nbins_foes      <- NB$nbins_foe
nbins_soe       <- NB$nbins_soe
}
