% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_indices.R
\name{sensitivity_indices}
\alias{sensitivity_indices}
\title{sensitivity_indices -}
\usage{
sensitivity_indices(output, inputs)
}
\arguments{
\item{output}{An array containing the target variable (Y).}

\item{inputs}{A two dimensional array of input variables (Xs).}
}
\value{
FOE - first-order effects (also called 'main' or 'individual' effects).

SOE - second-order effects (also called 'interaction' effects).

SI  - significance index, combined effect of each input.
}
\description{
sensitivity_indices calculates how much variability of the output is explained by inputs.
It uses the the number_of_bins, bin_data_1D, bin_data_2D  and magic_binning functions.
}
\examples{

library(Simdec)
data(example_data)
output                    <- example_data[,1]
inputs                    <- example_data[,2:5]

# Calculating sensitivity indices

S      <- sensitivity_indices(output, inputs)

}
