% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Significance.R
\name{significance}
\alias{significance}
\title{significance -}
\usage{
significance(output, inputs)
}
\arguments{
\item{output}{A vector containing the target variable (Y).}

\item{inputs}{All input variables.}
}
\value{
FOE - first-order effects (also called 'main' or 'individual' effects).

SOE - second-order effects (also called 'interaction' effects).

SI  - significance index, combined effect of each input.
}
\description{
significance calculates how much variability of the output is explained by inputs.
It uses the the number_of_bins, bin_data_1D, bin_data_2D  and magic_binning functions.
}
\examples{
df <- data.frame(Y  <- c(167, 82, 75, 134, 186, 51, 17, 167, 86, 198),
                 X1 <- c(10, 20, 30, 15, 8, 40, 60, 10, 17, 4),
                 X2 <- c(4,9,8,7,6,17,6,7,8,3))
output <- df[,1]
inputs <- df[,2:3]
S      <- significance(output, inputs)
}
