% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdec_visualization.R
\name{simdec_visualization}
\alias{simdec_visualization}
\title{simdec_visualization -}
\usage{
simdec_visualization(
  output,
  scenario,
  scenario_legend,
  main_colors,
  axistitle,
  var_names_dec
)
}
\arguments{
\item{output}{- an array of Y values}

\item{scenario}{- an array of associated scenario indices}

\item{scenario_legend}{- a scenario table that shows which states of which
variables compose different scenarios}

\item{main_colors}{- (optional) hex values of the main colors, should be as many as the
states for the first in decomposition (or the most
influential input variable). For example,
main_colors <- c('#00B0F0', '#1FDF4D', '#1FDF4D')}

\item{axistitle}{- a title for the x axis.}

\item{var_names_dec}{- cell array with names for variables for decomposition}
}
\value{
simdec_plot     - a stacked histogram

legend_table    - a table of legends
}
\description{
function visually represents results.
}
\examples{
BSC <- simdec_visualization(output, scenario, scenario_legend, main_colors, axistitle, var_names_dec)
}
