% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Significance.R
\name{weighted_variance}
\alias{weighted_variance}
\title{weighted_variance -}
\usage{
weighted_variance(x, weights)
}
\arguments{
\item{x}{- a numeric vector}

\item{weights}{- a vector of weights}
}
\value{
weighted variance
}
\description{
a function that calculates weighted variance.
}
\examples{

X       <- c(10, 20, 30, 15, 8, 40, 60, 10, 17, 4)
weights <- c(0.1, 0.3, 0.05, 0.05, 0.02, 0.08, 0.1, 0.1, 0.15, 0.05)
WV      <- weighted_variance(X, weights)
}
